+++ src/main/java/com/nisovin/shopkeepers/commands/shopkeepers/CommandList.java
@@ -36,10 +36,11 @@ import com.nisovin.shopkeepers.util.TextUtils;
 
 class CommandList extends Command {
 
+	private static final String ARGUMENT_ALL = "all";
+	private static final String ARGUMENT_ADMIN = "admin";
 	private static final String ARGUMENT_PLAYER = "player";
 	private static final String ARGUMENT_PLAYER_NAME = "player:name";
 	private static final String ARGUMENT_PLAYER_UUID = "player:uuid";
-	private static final String ARGUMENT_ADMIN = "admin";
 	private static final String ARGUMENT_PAGE = "page";
 
 	private static final int ENTRIES_PER_PAGE = 8;
@@ -57,6 +58,7 @@ class CommandList extends Command {
 
 		// arguments:
 		this.addArgument(new FirstOfArgument("target", Arrays.asList(
+				new LiteralArgument(ARGUMENT_ALL),
 				new LiteralArgument(ARGUMENT_ADMIN),
 				new FirstOfArgument(ARGUMENT_PLAYER, Arrays.asList(
 						// TODO provide completions for known shop owners?
@@ -88,13 +90,20 @@ class CommandList extends Command {
 	protected void execute(CommandInput input, CommandContextView context) throws CommandException {
 		CommandSender sender = input.getSender();
 		int page = context.get(ARGUMENT_PAGE);
+		boolean listAllShops = context.has(ARGUMENT_ALL); // can be null
 		boolean listAdminShops = context.has(ARGUMENT_ADMIN); // can be null
 		UUID targetPlayerUUID = context.get(ARGUMENT_PLAYER_UUID); // can be null
-		String targetPlayerName = context.get(ARGUMENT_PLAYER_NAME);
-		assert listAdminShops ^ (targetPlayerUUID != null ^ targetPlayerName != null); // xor
+		String targetPlayerName = context.get(ARGUMENT_PLAYER_NAME); // can be null
+		assert listAllShops ^ listAdminShops ^ (targetPlayerUUID != null ^ targetPlayerName != null); // xor
 
 		List<? extends Shopkeeper> shops;
-		if (listAdminShops) {
+		if (listAllShops) {
+			// permission check:
+			this.checkPermission(sender, ShopkeepersPlugin.LIST_ADMIN_PERMISSION);
+			this.checkPermission(sender, ShopkeepersPlugin.LIST_OTHERS_PERMISSION);
+	
+			shops = new ArrayList<>(shopkeeperRegistry.getAllShopkeepers());
+		} else if (listAdminShops) {
 			// permission check:
 			this.checkPermission(sender, ShopkeepersPlugin.LIST_ADMIN_PERMISSION);
 
@@ -160,7 +169,14 @@ class CommandList extends Command {
 		int maxPage = Math.max(1, (int) Math.ceil((double) shopsCount / ENTRIES_PER_PAGE));
 		page = Math.max(1, Math.min(page, maxPage));
 
-		if (listAdminShops) {
+		if (listAllShops) {
+			// listing all shops:
+			TextUtils.sendMessage(sender, Settings.msgListAllShopsHeader,
+					"shopsCount", shopsCount,
+					"page", page,
+					"maxPage", maxPage
+			);
+		} else if (listAdminShops) {
 			// listing admin shops:
 			TextUtils.sendMessage(sender, Settings.msgListAdminShopsHeader,
 					"shopsCount", shopsCount,